+++ src/main/java/com/grow/member_service/achievement/accomplished/application/event/AchievementAchievedEvent.java
@@ -1,0 +1,13 @@
+package com.grow.member_service.achievement.accomplished.application.event;
+
+import java.time.LocalDateTime;
+
+public record AchievementAchievedEvent(
+	Long accomplishedId,
+	Long memberId,
+	Long challengeId,
+	String challengeName,
+	int rewardPoint,
+	LocalDateTime occurredAt,
+	String dedupKey
+) {}
