+++ src/main/java/com/grow/member_service/achievement/accomplished/application/event/AchievementEventPublisher.java
@@ -1,0 +1,5 @@
+package com.grow.member_service.achievement.accomplished.application.event;
+
+public interface AchievementEventPublisher {
+	void publish(AchievementAchievedEvent e);
+}
