+++ src/main/java/com/grow/member_service/achievement/challenge/domain/repository/ChallengeRepository.java
@@ -1,0 +1,9 @@
+package com.grow.member_service.achievement.challenge.domain.repository;
+
+import java.util.Optional;
+
+import com.grow.member_service.achievement.challenge.domain.model.Challenge;
+
+public interface ChallengeRepository {
+	Optional<Challenge> findById(Long id);
+}
