+++ src/main/java/com/grow/member_service/achievement/challenge/infra/persistence/repository/ChallengeJpaRepository.java
@@ -1,0 +1,8 @@
+package com.grow.member_service.achievement.challenge.infra.persistence.repository;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import com.grow.member_service.achievement.challenge.infra.persistence.entity.ChallengeJpaEntity;
+
+public interface ChallengeJpaRepository extends JpaRepository<ChallengeJpaEntity, Long> {
+}
