--- src/main/java/com/grow/member_service/challenge/accomplished/application/dto/AccomplishedResponse.java
@@ -1,30 +1,0 @@
-package com.grow.member_service.challenge.accomplished.application.dto;
-
-import java.time.LocalDateTime;
-
-import com.grow.member_service.challenge.accomplished.domain.model.Accomplished;
-
-import lombok.AccessLevel;
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Getter
-@Builder
-@AllArgsConstructor
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-public class AccomplishedResponse {
-
-	private Long accomplishedId;
-	private Long challengeId;
-	private LocalDateTime accomplishedAt;
-
-	public static AccomplishedResponse from(Accomplished d) {
-		return AccomplishedResponse.builder()
-			.accomplishedId(d.getAccomplishedId())
-			.challengeId(d.getChallengeId())
-			.accomplishedAt(d.getAccomplishedAt())
-			.build();
-	}
-}
