--- src/main/java/com/grow/member_service/challenge/accomplished/application/dto/CreateAccomplishedRequest.java
@@ -1,15 +1,0 @@
-package com.grow.member_service.challenge.accomplished.application.dto;
-
-import jakarta.validation.constraints.NotNull;
-import lombok.AccessLevel;
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Getter
-@AllArgsConstructor
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-public class CreateAccomplishedRequest {
-	@NotNull
-	private Long challengeId;
-}
