--- src/main/java/com/grow/member_service/challenge/accomplished/application/event/AchievementEventPublisher.java
@@ -1,5 +1,0 @@
-package com.grow.member_service.challenge.accomplished.application.event;
-
-public interface AchievementEventPublisher {
-	void publish(AchievementAchievedEvent e);
-}
